+++ crates/chia-bls/benches/parse.rs
@@ -1,7 +1,6 @@
-use chia_bls::secret_key::SecretKey;
-use chia_bls::signature::sign;
-use chia_bls::PublicKey;
+use chia_bls::sign;
 use chia_bls::Signature;
+use chia_bls::{PublicKey, SecretKey};
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
