+++ crates/chia-bls/fuzz/fuzz_targets/derive.rs
@@ -1,10 +1,8 @@
 #![no_main]
 use libfuzzer_sys::fuzz_target;
 
-use chia_bls::derivable_key::DerivableKey;
-use chia_bls::public_key::PublicKey;
-use chia_bls::secret_key::SecretKey;
-use chia_bls::signature::{sign, verify};
+use chia_bls::{sign, verify};
+use chia_bls::{DerivableKey, PublicKey, SecretKey};
 
 fuzz_target!(|data: &[u8]| {
     if data.len() < 32 {
