+++ crates/chia-bls/src/public_key.rs
@@ -9,23 +9,10 @@ use std::io::Cursor;
 use std::mem::MaybeUninit;
 use std::ops::{Add, AddAssign, Neg, SubAssign};
 
-#[cfg(feature = "py-bindings")]
-use crate::{GTElement, Signature};
-#[cfg(feature = "py-bindings")]
-use chia_py_streamable_macro::PyStreamable;
-#[cfg(feature = "py-bindings")]
-use chia_traits::from_json_dict::FromJsonDict;
-#[cfg(feature = "py-bindings")]
-use chia_traits::to_json_dict::ToJsonDict;
-#[cfg(feature = "py-bindings")]
-use pyo3::prelude::PyAnyMethods;
-#[cfg(feature = "py-bindings")]
-use pyo3::{pyclass, pymethods, IntoPy, PyAny, PyObject, PyResult, Python};
-
 #[cfg_attr(
     feature = "py-bindings",
-    pyclass(name = "G1Element"),
-    derive(PyStreamable)
+    pyo3::pyclass(name = "G1Element"),
+    derive(chia_py_streamable_macro::PyStreamable)
 )]
 #[derive(Clone, Copy, Default)]
 pub struct PublicKey(pub(crate) blst_p1);
@@ -160,46 +147,6 @@ impl PublicKey {
     }
 }
 
-#[cfg(feature = "py-bindings")]
-#[pymethods]
-impl PublicKey {
-    #[classattr]
-    const SIZE: usize = 48;
-
-    #[new]
-    pub fn init() -> Self {
-        Self::default()
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "generator")]
-    pub fn py_generator() -> Self {
-        Self::generator()
-    }
-
-    pub fn pair(&self, other: &Signature) -> GTElement {
-        other.pair(self)
-    }
-
-    #[pyo3(name = "get_fingerprint")]
-    pub fn py_get_fingerprint(&self) -> u32 {
-        self.get_fingerprint()
-    }
-
-    fn __str__(&self) -> String {
-        hex::encode(self.to_bytes())
-    }
-
-    #[must_use]
-    pub fn __add__(&self, rhs: &Self) -> Self {
-        self + rhs
-    }
-
-    pub fn __iadd__(&mut self, rhs: &Self) {
-        *self += rhs;
-    }
-}
-
 impl PartialEq for PublicKey {
     fn eq(&self, other: &Self) -> bool {
         unsafe { blst_p1_is_equal(&self.0, &other.0) }
@@ -299,69 +246,6 @@ impl fmt::Debug for PublicKey {
     }
 }
 
-#[cfg(feature = "py-bindings")]
-impl ToJsonDict for PublicKey {
-    fn to_json_dict(&self, py: Python<'_>) -> PyResult<PyObject> {
-        let bytes = self.to_bytes();
-        Ok(("0x".to_string() + &hex::encode(bytes)).into_py(py))
-    }
-}
-
-#[cfg(feature = "py-bindings")]
-pub fn parse_hex_string(o: &pyo3::Bound<'_, PyAny>, len: usize, name: &str) -> PyResult<Vec<u8>> {
-    use pyo3::exceptions::{PyTypeError, PyValueError};
-    if let Ok(s) = o.extract::<String>() {
-        let s = if let Some(st) = s.strip_prefix("0x") {
-            st
-        } else {
-            &s[..]
-        };
-        let buf = match hex::decode(s) {
-            Err(_) => {
-                return Err(PyValueError::new_err("invalid hex"));
-            }
-            Ok(v) => v,
-        };
-        if buf.len() == len {
-            Ok(buf)
-        } else {
-            Err(PyValueError::new_err(format!(
-                "{}, invalid length {} expected {}",
-                name,
-                buf.len(),
-                len
-            )))
-        }
-    } else if let Ok(buf) = o.extract::<Vec<u8>>() {
-        if buf.len() == len {
-            Ok(buf)
-        } else {
-            Err(PyValueError::new_err(format!(
-                "{}, invalid length {} expected {}",
-                name,
-                buf.len(),
-                len
-            )))
-        }
-    } else {
-        Err(PyTypeError::new_err(format!(
-            "invalid input type for {name}"
-        )))
-    }
-}
-
-#[cfg(feature = "py-bindings")]
-impl FromJsonDict for PublicKey {
-    fn from_json_dict(o: &pyo3::Bound<'_, PyAny>) -> PyResult<Self> {
-        Ok(Self::from_bytes(
-            parse_hex_string(o, 48, "PublicKey")?
-                .as_slice()
-                .try_into()
-                .unwrap(),
-        )?)
-    }
-}
-
 impl DerivableKey for PublicKey {
     fn derive_unhardened(&self, idx: u32) -> Self {
         let mut hasher = Sha256::new();
@@ -409,6 +293,73 @@ pub fn hash_to_g1_with_dst(msg: &[u8], dst: &[u8]) -> PublicKey {
         p1.assume_init()
     };
     PublicKey(p1)
+}
+
+#[cfg(feature = "py-bindings")]
+mod pybindings {
+    use super::*;
+
+    use crate::{parse_hex::parse_hex_string, GTElement, Signature};
+
+    use chia_traits::{FromJsonDict, ToJsonDict};
+    use pyo3::prelude::*;
+
+    #[pymethods]
+    impl PublicKey {
+        #[classattr]
+        const SIZE: usize = 48;
+
+        #[new]
+        pub fn init() -> Self {
+            Self::default()
+        }
+
+        #[staticmethod]
+        #[pyo3(name = "generator")]
+        pub fn py_generator() -> Self {
+            Self::generator()
+        }
+
+        pub fn pair(&self, other: &Signature) -> GTElement {
+            other.pair(self)
+        }
+
+        #[pyo3(name = "get_fingerprint")]
+        pub fn py_get_fingerprint(&self) -> u32 {
+            self.get_fingerprint()
+        }
+
+        fn __str__(&self) -> String {
+            hex::encode(self.to_bytes())
+        }
+
+        #[must_use]
+        pub fn __add__(&self, rhs: &Self) -> Self {
+            self + rhs
+        }
+
+        pub fn __iadd__(&mut self, rhs: &Self) {
+            *self += rhs;
+        }
+    }
+
+    impl ToJsonDict for PublicKey {
+        fn to_json_dict(&self, py: Python<'_>) -> PyResult<PyObject> {
+            let bytes = self.to_bytes();
+            Ok(("0x".to_string() + &hex::encode(bytes)).into_py(py))
+        }
+    }
+
+    impl FromJsonDict for PublicKey {
+        fn from_json_dict(o: &Bound<'_, PyAny>) -> PyResult<Self> {
+            Ok(Self::from_bytes(
+                parse_hex_string(o, 48, "PublicKey")?
+                    .as_slice()
+                    .try_into()
+                    .unwrap(),
+            )?)
+        }
+    }
 }
 
 #[cfg(test)]
@@ -758,6 +709,7 @@ mod tests {
 mod pytests {
     use super::*;
     use crate::SecretKey;
+    use pyo3::{IntoPy, Python};
     use rand::rngs::StdRng;
     use rand::{Rng, SeedableRng};
     use rstest::rstest;
